IMMORTAL: Fix Various GCC Compiler Warnings
Some changes to this code are not functional, but sufficient to fix
the warnings i.e. adding of fallthrough comments, removal of extraneous
semi-colons. However, some of these changes will require analysis and
playtesting to check that these are correct i.e. fixes to the size of
structure variables, additional of breaks in some switch statements.

+++ engines/immortal/story.cpp
@@ -79,8 +79,8 @@ void ImmortalEngine::initStoryDynamic() {
 	 * including spawn point and entry/exit points
 	 */ 
 	int univRoom = 4;               // The room the player starts in when beginning this level
-	uint8 univRoomX = 512;
-	uint8 univRoomY = 416;
+	uint16 univRoomX = 512;
+	uint16 univRoomY = 416;
 
 	_stories[0]._level = 0;
 	_stories[0]._part  = 1;
